requirements.txt
streamlit>=1.20.0
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.10.0
python-dateutil>=2.8.2